refactor component logic into hook
